+++ test/projects/standard/server/lifecycle/on-render-after-data-to-store.js
@@ -1,0 +1,4 @@
+export default async (/*o*/) => {
+    if (__DEV__) console.log('server lifecycle: onRenderAfterDataToStore')
+    // if (__DEV__) console.log(o)
+}